+++ pkg/pt/material.go
@@ -11,25 +11,24 @@ type scatterResult struct {
 }
 
 type Material interface {
-	scatter(ray, *hit, *rand.Rand) (bool, *scatterResult)
+	scatter(ray, *hit, *rand.Rand) (bool, scatterResult)
 }
 
 type Diffuse struct {
 	Albedo Color
 }
 
-func (d *Diffuse) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, *scatterResult) {
+func (d *Diffuse) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, scatterResult) {
 	// TODO: Revise random ray generation
 	scatterDirection := intersec.normal.Add(RandomUnitVector(r))
 	if scatterDirection.ApproxZero() {
 		scatterDirection = intersec.normal
 	}
 	ray.reuse(intersec.point, scatterDirection)
-	result := &scatterResult{
+	return true, scatterResult{
 		scattered:   ray,
 		attenuation: d.Albedo,
 	}
-	return true, result
 }
 
 type Reflective struct {
@@ -37,14 +36,13 @@ type Reflective struct {
 	Diffusion float64 // diffusion in range [0,1]
 }
 
-func (d *Reflective) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, *scatterResult) {
+func (d *Reflective) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, scatterResult) {
 	reflected := reflect(ray.direction.Unit(), intersec.normal)
 	ray.reuse(intersec.point, reflected.Add(RandomUnitVector(r).Mul(d.Diffusion)))
-	result := &scatterResult{
+	return reflected.Dot(intersec.normal) > 0, scatterResult{
 		scattered:   ray,
 		attenuation: d.Albedo,
 	}
-	return reflected.Dot(intersec.normal) > 0, result
 }
 
 type Refractive struct {
@@ -52,7 +50,7 @@ type Refractive struct {
 	Ratio  float64
 }
 
-func (d *Refractive) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, *scatterResult) {
+func (d *Refractive) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, scatterResult) {
 	defractionRatio := d.Ratio
 	if intersec.frontFace {
 		defractionRatio = 1 / d.Ratio
@@ -71,7 +69,7 @@ func (d *Refractive) scatter(ray ray, intersec *hit, r *rand.Rand) (bool, *scatt
 		direction = refract(unitDir, intersec.normal, defractionRatio)
 	}
 	ray.reuse(intersec.point, direction)
-	return true, &scatterResult{
+	return true, scatterResult{
 		scattered:   ray,
 		attenuation: d.Albedo,
 	}