+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Using Ray call by value insted of pointer for perfomance increase
 - Reusing Rays when scattering and casting new rays from camera for performance increase
+- Return scatter result as struct for performance increase
 
 ## [0.0.1] - 2021-07-15
 ### Added